score integration to gameloop
