+++ frontend/src/components/Game/ResultButton.js
@@ -1,8 +1,26 @@
-import React, { useState } from 'react'
+import { useState, useEffect } from "react"; 
 
 
 export default function ResultButton() {
-     const [score, setScore] = useState(0);
+    const [scores, setScores] = useState([]);
+    const [score, setScore] = useState(0);
+
+    //Load scores from local storage on pageload
+    useEffect(() => {
+            const json = localStorage.getItem("scores");
+            const savedScores = JSON.parse(json);
+            if (savedScores) { setScores(savedScores); }
+        }, 
+        []
+    );
+
+    //Save scores to local storage when changed
+    useEffect(() => { 
+        const json = JSON.stringify(scores);
+        localStorage.setItem("scores", json);
+        }, 
+        [scores]
+    );
 
     const compare = () => {
 
@@ -39,8 +57,15 @@ export default function ResultButton() {
                 document.getElementById(boxes[i]).style.backgroundColor = "red";
             }
         }
+        
+        const newScore = {
+            //Potentially can add more info here; eg date, difficulty
+            id: Math.random().toString(36),
+            text: Score,
+        };
+        setScores([...scores, newScore]);
         setScore(Score);
-        //console.log(answer);
+
     }
